argmin_testfunctions

Function rastrigin_derivative_const

pub fn rastrigin_derivative_const<const N: usize, T>(param: &[T; N]) -> [T; N]
where
    T: Float + FromPrimitive + Sum + Into<f64>,

Derivative of Rastrigin test function with a = 10.

This is the const generics version, which requires the number of parameters to be known at compile time.