Crate argmin_testfunctions

Test functions for optimization algorithms

Functions

ackley

Ackley test function

ackley_abc

Ackley test function

beale

Beale test function

beale_derivative

Derivative of Beale test function

beale_hessian

Derivative of Beale test function

booth

Booth test function

booth_derivative

Derivative of Booth test function

booth_hessian

Hessian of Booth test function

bukin_n6

Bukin test function No. 6

bukin_n6_derivative

Derivative of Bukin test function No. 6

bukin_n6_hessian

Hessian of Bukin test function No. 6

cross_in_tray

Cross-in-tray test function

easom

Easom test function

easom_derivative

Derivative of Easom test function

easom_hessian

Hessian of Easom test function

eggholder

Eggholder test function

goldsteinprice

Goldstein-Price test function

himmelblau

Himmelblau test function

holder_table

Holder table test function

levy

Levy test function

levy_n13

Levy test function No. 13

matyas

Matyas test function

matyas_derivative

Derivative of Matyas test function

matyas_hessian

Hessian of Matyas test function

mccorminck

McCorminck test function

picheny

Picheny test function

rastrigin

Rastrigin test function

rastrigin_a

Rastrigin test function

rastrigin_a_derivative

Derivative of Rastrigin test function where a can be chosen freely

rastrigin_a_derivative_const

Derivative of Rastrigin test function where a can be chosen freely

rastrigin_a_hessian

Hessian of Rastrigin test function where a can be chosen freely

rastrigin_a_hessian_const

Hessian of Rastrigin test function where a can be chosen freely

rastrigin_derivative

Derivative of Rastrigin test function

rastrigin_derivative_const

Derivative of Rastrigin test function

rastrigin_hessian

Hessian of Rastrigin test function

rastrigin_hessian_const

Hessian of Rastrigin test function

rosenbrock

Multidimensional Rosenbrock test function

rosenbrock_derivative

Derivative of the multidimensional Rosenbrock test function

rosenbrock_derivative_const

Derivative of the multidimensional Rosenbrock test function

rosenbrock_hessian

Hessian of the multidimensional Rosenbrock test function

rosenbrock_hessian_const

Hessian of the multidimensional Rosenbrock test function

schaffer_n2

Schaffer test function No. 2

schaffer_n4

Schaffer test function No. 4

sphere

Sphere test function

sphere_derivative

Derivative of sphere test function

sphere_derivative_const

Derivative of sphere test function

sphere_hessian

Hessian of sphere test function

sphere_hessian_const

Hessian of sphere test function

styblinski_tang

Styblinski-Tang test function

threehumpcamel

Three-hump camel test function

zero

Zero test function

zero_derivative

Derivative of zero test function